+++ src/Providers/Anthropic/Parsers/StreamEventParser.php
@@ -1,0 +1,121 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Prism\Prism\Providers\Anthropic\Parsers;
+
+use Prism\Prism\Exceptions\PrismChunkDecodeException;
+use Psr\Http\Message\StreamInterface;
+use Throwable;
+
+class StreamEventParser
+{
+    /**
+     * @return array<string, mixed>|null
+     */
+    public function parseNextChunk(StreamInterface $stream): ?array
+    {
+        $line = $this->readLine($stream);
+        $line = trim($line);
+
+        if ($line === '' || $line === '0') {
+            return null;
+        }
+
+        if (str_starts_with($line, 'event:')) {
+            return $this->parseEventChunk($line, $stream);
+        }
+
+        if (str_starts_with($line, 'data:')) {
+            return $this->parseDataChunk($line);
+        }
+
+        return null;
+    }
+
+    /**
+     * @return array<string, mixed>|null
+     */
+    protected function parseEventChunk(string $line, StreamInterface $stream): ?array
+    {
+        $eventType = trim(substr($line, strlen('event:')));
+
+        if ($eventType === 'ping') {
+            return ['type' => 'ping'];
+        }
+
+        $dataLine = $this->readLine($stream);
+        $dataLine = trim($dataLine);
+
+        if ($dataLine === '' || $dataLine === '0') {
+            return ['type' => $eventType];
+        }
+
+        if (! str_starts_with($dataLine, 'data:')) {
+            return ['type' => $eventType];
+        }
+
+        return $this->parseJsonData($dataLine, $eventType);
+    }
+
+    /**
+     * @return array<string, mixed>|null
+     */
+    protected function parseDataChunk(string $line): ?array
+    {
+        $jsonData = trim(substr($line, strlen('data:')));
+
+        if ($jsonData === '' || $jsonData === '0' || str_contains($jsonData, 'DONE')) {
+            return null;
+        }
+
+        return $this->parseJsonData($jsonData);
+    }
+
+    /**
+     * @return array<string, mixed>|null
+     */
+    protected function parseJsonData(string $jsonDataLine, ?string $eventType = null): ?array
+    {
+        $jsonData = trim(str_starts_with($jsonDataLine, 'data:')
+            ? substr($jsonDataLine, strlen('data:'))
+            : $jsonDataLine);
+
+        if ($jsonData === '' || $jsonData === '0') {
+            return $eventType ? ['type' => $eventType] : null;
+        }
+
+        try {
+            $data = json_decode($jsonData, true, flags: JSON_THROW_ON_ERROR);
+
+            if ($eventType) {
+                $data['type'] = $eventType;
+            }
+
+            return $data;
+        } catch (Throwable $e) {
+            throw new PrismChunkDecodeException('Anthropic', $e);
+        }
+    }
+
+    protected function readLine(StreamInterface $stream): string
+    {
+        $buffer = '';
+
+        while (! $stream->eof()) {
+            $byte = $stream->read(1);
+
+            if ($byte === '') {
+                return $buffer;
+            }
+
+            $buffer .= $byte;
+
+            if ($byte === "\n") {
+                break;
+            }
+        }
+
+        return $buffer;
+    }
+}
